requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.6
psycopg[binary]==3.2.4
ollama==0.4.7
pdfplumber==0.11.4
streamlit==1.41.1
pandas==2.2.3
plotly==5.24.1
